break out UserPage to work with storybook
