+++ .storybook/main.js
@@ -17,6 +17,11 @@ const config = {
     ...config,
     STORYBOOK_RUN: true,
   }),
+  webpackFinal: (config) => {
+    config.resolve.fallback.fs = false;
+
+    return config;
+  },
 };
 
 export default config
