+++ stories/components/user/UserPage.stories.js
@@ -1,5 +1,5 @@
-import User from "../../pages/[username]";
-import profile from "../../data/eddiejaoude.json";
+import User from "../../../components/user/UserPage";
+import profile from "../../../data/eddiejaoude.json";
 
 export default {
   component: User,
@@ -19,6 +19,6 @@ export const Basic = {
       views: 1,
       username: "eddiejaoude"
     },
-    BASE_URL: "https://linkfree.eddiehub.io/"
+    BASE_URL: "https://linkfree.eddiehub.io"
   }
 };
